+++ de.metas.purchasecandidate.base/src/main/java/de/metas/purchasecandidate/grossprofit/PurchaseProfitInfo.java
@@ -101,6 +101,5 @@ public class PurchaseProfitInfo
 			this.salesNetPrice = salesNetPrice;
 			return this;
 		}
-
 	}
 }
